+++ src/app/rootReducer.js
@@ -1,4 +1,5 @@
 import infoAlertReducer from '../features/alert/infoAlertSlice';
+import boardReducer from '../features/board/boardSlice';
 import acceptDrawDialogReducer from '../features/dialog/acceptDrawDialogSlice';
 import acceptRematchDialogReducer from '../features/dialog/acceptRematchDialogSlice';
 import acceptResignDialogReducer from '../features/dialog/acceptResignDialogSlice';
@@ -28,7 +29,6 @@ import modeReducer from '../features/mode/modeSlice';
 import gameTableReducer from '../features/table/gameTableSlice';
 import openingAnalysisTableReducer from '../features/table/openingAnalysisTableSlice';
 import variantReducer from '../features/variant/variantSlice';
-import boardReducer from '../features/boardSlice';
 import heuristicsBarReducer from '../features/heuristicsBarSlice';
 import historyReducer from '../features/historySlice';
 import mainButtonsReducer from '../features/mainButtonsSlice';
@@ -36,6 +36,7 @@ import serverReducer from '../features/wsSlice';
 
 const rootReducer = {
   infoAlert: infoAlertReducer,
+  board: boardReducer,
   acceptDrawDialog: acceptDrawDialogReducer,
   acceptRematchDialog: acceptRematchDialogReducer,
   acceptResignDialog: acceptResignDialogReducer,
@@ -63,7 +64,6 @@ const rootReducer = {
   watchDialog: watchDialogReducer,
   gameTable: gameTableReducer,
   openingAnalysisTable: openingAnalysisTableReducer,
-  board: boardReducer,
   heuristicsBar: heuristicsBarReducer,
   history: historyReducer,
   mainButtons: mainButtonsReducer,
