+++ src/common/Dispatcher.js
@@ -1,7 +1,7 @@
 import Opening from '../common/Opening.js';
-import * as board from '../features/boardSlice';
 import * as heuristicsBar from '../features/heuristicsBarSlice';
 import * as history from '../features/historySlice';
+import * as board from '../features/board/boardSlice';
 import * as infoAlert from '../features/alert/infoAlertSlice';
 import * as gameTable from '../features/table/gameTableSlice';
 import * as openingAnalysisTable from '../features/table/openingAnalysisTableSlice';
