+++ src/ws/WsEvent.js
@@ -2,9 +2,9 @@ import jwt_decode from "jwt-decode";
 import store from '../app/store';
 import Pgn from '../common/Pgn';
 import Dispatcher from '../common/Dispatcher';
-import * as board from '../features/boardSlice';
 import * as heuristicsBar from '../features/heuristicsBarSlice';
 import * as infoAlert from '../features/alert/infoAlertSlice';
+import * as board from '../features/board/boardSlice';
 import * as acceptDrawDialog from '../features/dialog/acceptDrawDialogSlice';
 import * as acceptRematchDialog from '../features/dialog/acceptRematchDialogSlice';
 import * as acceptTakebackDialog from '../features/dialog/acceptTakebackDialogSlice';
