+++ src/ws/WsEventListener.js
@@ -1,6 +1,6 @@
 import store from '../app/store';
 import Wording from '../common/Wording.js';
-import * as board from '../features/boardSlice';
+import * as board from '../features/board/boardSlice';
 import * as progressDialog from '../features/dialog/progressDialogSlice';
 import * as modeConst from '../features/mode/modeConst';
 import WsEvent from './WsEvent';
